loggers added.

+++ src/main/java/com/calculator/Operations.java
@@ -2,7 +2,7 @@ package com.calculator;
 
 public interface Operations {
     public double sqrt(double number);
-    public long factorial(int number);
+    public double factorial(int number);
     public double log(double number);
     public double power(double number, double exponent);
 }
